Sat, Oct 21, 2023, 4:57 PM +05:30

+++ half/half.c
@@ -18,9 +18,9 @@ int main(void)
 // TODO: Complete the function
 float half(float bill, float tax, int tip)
 {
-    tax = (float)(tax*bill)/100;
-    float new_total = (float)tax+ bill;
-    tip =(float) (tip*new_total)/100;
+    tax = (double)(tax*bill)/100;
+    float new_total = (double)(tax+ bill);
+    tip =(double) (tip*new_total)/100;
     float half_amount = (bill+tax+tip)/2;
     return half_amount;
